+++ plugins/woocommerce-admin/packages/experimental/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+-   Fix animations for collapsible list. #7429
+
 # 1.5.0
 
 -   Fix commonjs module build, allow package to be built in isolation. #7286